add TORCH_CUDA_CPP_API for AutoNcclGroup
torch::cuda::nccl is an option to depend only on torch but not nccl. but to use torch::cuda::nccl::send/torch::cuda::nccl::recv, ncclGroupStart()/ncclGroupEnd() is needed,  torch::cuda::nccl::AutoNcclGroup can be used.  but torch::cuda::nccl::AutoNcclGroup is not exported and is LOCAL symbol, which can't be used from outside of libtorch.