Switch batch norm stack to consolidated ops
Summary: This commit switches `aten.batch_norm` to call the
new `batch_norm_with_update` and `batch_norm_no_update` ops,
instead of the old `_batch_norm_impl_index` op. The new stack
is "consolidated" in the sense that there is a single backend
agnostic op that will internally pick the right kernel based
on the backend, but this detail will be hidden away from the
user and from the model graph.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 0000ba9b4ee7cf6edd40feebe8aa14c8c327d634
Pull Request resolved: https://github.com/pytorch/pytorch/pull/119496